refactor references to internal/limitrange to pkg/limitrange
