+++ internal/mutators/podtemplatespec.go
@@ -3,8 +3,8 @@ package mutators
 import (
 	"fmt"
 
-	"github.com/kanopy-platform/hedgetrimmer/internal/limitrange"
 	"github.com/kanopy-platform/hedgetrimmer/internal/quantity"
+	"github.com/kanopy-platform/hedgetrimmer/pkg/limitrange"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 )
